requirements.txt
altgraph==0.17.4
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Faker==36.1.1
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
packaging==23.2
pefile==2023.2.7
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
pywin32-ctypes==0.2.2
setuptools==69.0.3
tzdata==2025.1
Werkzeug==3.1.3
